Add Abstract Payload
Allows extending abstract implementation without forcing Status constants
