+++ src/Payload.php
@@ -3,7 +3,7 @@ namespace Aura\Payload;
 
 use Aura\Payload_Interface\PayloadInterface;
 
-class Payload implements PayloadInterface
+class Payload extends AbstractPayload
 {
     const ACCEPTED = 'ACCEPTED';
     const AUTHENTICATED = 'AUTHENTICATED';
@@ -26,64 +26,4 @@ class Payload implements PayloadInterface
     const UPDATED = 'UPDATED';
     const VALID = 'VALID';
 
-    protected $status;
-    protected $input;
-    protected $output;
-    protected $messages;
-    protected $extras;
-
-    public function setStatus($status)
-    {
-        $this->status = $status;
-        return $this;
-    }
-
-    public function getStatus()
-    {
-        return $this->status;
-    }
-
-    public function setInput($input)
-    {
-        $this->input = $input;
-        return $this;
-    }
-
-    public function getInput()
-    {
-        return $this->input;
-    }
-
-    public function setOutput($output)
-    {
-        $this->output = $output;
-        return $this;
-    }
-
-    public function getOutput()
-    {
-        return $this->output;
-    }
-
-    public function setMessages($messages)
-    {
-        $this->messages = $messages;
-        return $this;
-    }
-
-    public function getMessages()
-    {
-        return $this->messages;
-    }
-
-    public function setExtras($extras)
-    {
-        $this->extras = $extras;
-        return $this;
-    }
-
-    public function getExtras()
-    {
-        return $this->extras;
-    }
 }
